updated team section
